Add branch analysis pass for MiGPU
Current implementation checks for the first terminator in each BB to see
if its a conditional branch or a simple fall through. Two sets of
instructions will be inserted before this branch instruction:
1. BNE, BEQ
```
	%tmp = SUB %op1, %op2
	%tmp2 = SLTIU %tmp, 1
	SPLIT %tmp2
```
2. Others
```
	%tmp = SLT %op1, %op2
	SPLIT %tmp
```
The actual logic do not need to follow the orignal branch condition
since `split` implementation in the hardware only requires predicates to
split into to distinct groups, zero and non-zero.

`JOIN` will be inserted in the begining of the IPDOM of the successor of
the current diverged BB. A IPDOM is expected to exist.

There is no test, need help testing this pass.